Incorporated new wingpanel slim settings

+++ src/AppearanceGrid.vala
@@ -193,6 +193,7 @@ public class AppearanceGrid : Gtk.Grid
         var wingpanel = new Gtk.Switch ();
         var wingpanel_slim = CerbereSettings.get_default ().monitored_processes;
         var checkslim = File.new_for_path ("/usr/bin/wingpanel-slim");
+        var checkschema = File.new_for_path ("/usr/share/glib-2.0/schemas/org.pantheon.desktop.wingpanel-slim.gschema.xml");
         int pos = -1;
 
         for (int i = 0; i < wingpanel_slim.length ; i++) {
@@ -200,12 +201,41 @@ public class AppearanceGrid : Gtk.Grid
             pos = i; 
         }
 
-        if (checkslim.query_exists() && pos != -1){
+        if (checkschema.query_exists() && checkslim.query_exists() && pos != -1){
+
+            /* Wingpanel Position */
+            var slim_pos_box = new Gtk.Box (Gtk.Orientation.HORIZONTAL, 0);
+            var slim_pos = new Gtk.ComboBoxText ();
+            var slim_label = new LLabel.right (_("Slim Wingpanel Position:"));
+            slim_pos.append ("Right", _("Right"));
+            slim_pos.append ("Middle", _("Middle"));
+            slim_pos.append ("Left", _("Left"));
+
+            slim_pos.active_id = WingpanelslimSettings.get_default ().panel_position;
+            slim_pos.changed.connect (() => WingpanelslimSettings.get_default ().panel_position = slim_pos.active_id );
+            slim_pos.halign = Gtk.Align.START;
+            slim_pos.width_request = 160;
+
+            var slim_pos_default = new Gtk.ToolButton.from_stock (Gtk.Stock.REVERT_TO_SAVED);
+
+            slim_pos_default.clicked.connect (() => {
+                WingpanelslimSettings.get_default ().schema.reset ("panel-position");
+                slim_pos.active_id = WingpanelslimSettings.get_default ().panel_position;
+            });
+            slim_pos_default.halign = Gtk.Align.START;
+
+            slim_pos_box.pack_start (slim_pos, false);
+            slim_pos_box.pack_start (slim_pos_default, false);
+
+            this.attach (slim_label, 0, 7, 2, 1);
+            this.attach (slim_pos_box, 2, 7, 2, 1);
 
             if ( wingpanel_slim[pos] == "wingpanel-slim" )
                 wingpanel.set_active(true);
 
             wingpanel.notify["active"].connect (() => {
+                slim_label.set_sensitive(wingpanel.active);
+                slim_pos_box.set_sensitive(wingpanel.active);
                 (wingpanel.active)?wingpanel_slim[pos] = "killall wingpanel":wingpanel_slim[pos] = "killall wingpanel-slim";
                 CerbereSettings.get_default ().monitored_processes = wingpanel_slim;
                 (wingpanel.active)?wingpanel_slim[pos] = "wingpanel-slim":wingpanel_slim[pos] = "wingpanel";
@@ -215,7 +245,7 @@ public class AppearanceGrid : Gtk.Grid
             wingpanel.halign = Gtk.Align.START;
             this.attach (new LLabel.right (_("Slim Wingpanel:")), 0, 6, 2, 1);
             this.attach (wingpanel, 2, 6, 2, 1);
-           
+
         } else if (!checkslim.query_exists() && wingpanel_slim[pos] == "wingpanel-slim") {
             wingpanel_slim[pos] = "killall wingpanel";
             CerbereSettings.get_default ().monitored_processes = wingpanel_slim;
